fix: repair imports/ specifiers broken by Phase 4 moves (tsc-only breakage)
apps/meteor/imports/ is ignored by ESLint, so three stale imports of
app/authorization/server/functions/hasPermission in
imports/personal-access-tokens/ survived every lint pass and only
failed under yarn typecheck.

- rewrite the three specifiers to server/lib/authorization/hasPermission
- add imports/, client/, definition/, packages/ to the search dirs of
  move-module.mjs and verify-no-old-imports.mjs so later phases rewrite
  and verify those trees too
- document the lint blind spot in MIGRATION_PLAN.md: every phase must
  run yarn typecheck in addition to lint
