+++ apps/meteor/imports/personal-access-tokens/server/api/methods/generateToken.ts
@@ -4,7 +4,7 @@ import { Accounts } from 'meteor/accounts-base';
 import type { ServerMethods } from '@rocket.chat/ddp-client';
 import { Users } from '@rocket.chat/models';
 
-import { hasPermissionAsync } from '../../../../../app/authorization/server/functions/hasPermission';
+import { hasPermissionAsync } from '../../../../../server/lib/authorization/hasPermission';
 import { twoFactorRequired } from '../../../../../app/2fa/server/twoFactorRequired';
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
